refactor(phase-4): Split large files into focused modules
Task 4.1 - applescript_manager.py Split (COMPLETE):
- Reduced from 1,706 → 952 lines (44% reduction)
- Extracted executor.py (146 lines) - AppleScript execution with locking/retry
- Extracted formatters.py (409 lines) - Date/tag/URL formatting and parsing
- Extracted queries.py (353 lines) - AppleScript query builders
- applescript_manager.py now acts as a facade, delegating to specialized modules
- All 392 tests passing

Task 4.2 - tools.py Split (PARTIAL):
- Reduced from 1,696 → 1,660 lines (minimal reduction)
- Extracted tools_helpers/helpers.py (217 lines) - Utility conversion methods
- Created ToolsHelpers class with static methods for:
  - escape_applescript_string()
  - convert_to_boolean()
  - convert_applescript_todo() / convert_todo()
  - convert_project() / convert_area()
- Tools.py now delegates basic conversions to helpers module
- All 392 tests passing

File Organization:
- services/applescript/ now contains:
  - __init__.py (13 lines) - Exports all AppleScript classes
  - parser.py (453 lines) - State machine parser (pre-existing)
  - executor.py (146 lines) - NEW - Execution with locking
  - formatters.py (409 lines) - NEW - Formatting utilities
  - queries.py (353 lines) - NEW - Query builders
  - applescript_manager.py (952 lines) - Facade pattern

- tools_helpers/ now contains:
  - __init__.py (5 lines)
  - helpers.py (217 lines) - NEW - Conversion utilities

Backwards Compatibility:
- 100% maintained through facade pattern
- All public APIs unchanged
- Delegation methods preserve original signatures
- No breaking changes
