+++ src/things_mcp/services/applescript/__init__.py
@@ -1,5 +1,13 @@
 """AppleScript parsing and processing utilities."""
 
 from .parser import AppleScriptParser
+from .executor import AppleScriptExecutor
+from .formatters import AppleScriptFormatters
+from .queries import AppleScriptQueries
 
-__all__ = ['AppleScriptParser']
+__all__ = [
+    'AppleScriptParser',
+    'AppleScriptExecutor',
+    'AppleScriptFormatters',
+    'AppleScriptQueries',
+]
