+++ src/things_mcp/tools_helpers/__init__.py
@@ -1,0 +1,5 @@
+"""Tools package for Things 3 MCP server."""
+
+from .helpers import ToolsHelpers
+
+__all__ = ['ToolsHelpers']
